Made few changes/modifications summarised below:
Integrated the core game logic with the persistence entities.

Made the MemoryGameExceptions in the package
mjb.memorygame.game.exceptions checked exceptions.

Abstracted some entity / integration code to
mjb.memorygame.services.GameServiceImpl

Added lock to the signature of the
MemoryGame(List<Integer> cards, List<Integer> moves, boolean lock, int firstPlayer)
constructor. This is necessary because in order to reproduce
the required state of a game at the end of a turn the value
of lock tells us whether the completeTurn() method has been
called yet: if false completeTurn() must be called.

+++ src/main/java/mjb/memorygame/entities/Move.java
@@ -13,11 +13,11 @@ public class Move {
     private long id;
 
     @Column
-    private long cardIndex;
+    private int cardIndex;
 
     public Move() {}
     
-    public Move(long cardIndex) {
+    public Move(int cardIndex) {
         setCardIndex(cardIndex);
     }
 
@@ -29,11 +29,11 @@ public class Move {
         this.id = id;
     }
 
-    public long getCardIndex() {
+    public int getCardIndex() {
         return cardIndex;
     }
 
-    public void setCardIndex(long cardIndex) {
+    public void setCardIndex(int cardIndex) {
         this.cardIndex = cardIndex;
     }
